Create seperate continuous mode file
